+++ FAnsiSql/Connections/IManagedConnection.cs
@@ -23,5 +23,11 @@ namespace FAnsi.Connections
         /// Optional - transaction being run (See <see cref="DiscoveredServer.BeginNewTransactedConnection"/>.  If this is not null then <see cref="Transaction"/> should also be not null.
         /// </summary>
         IManagedTransaction ManagedTransaction { get; }
+
+        /// <summary>
+        /// True to close the connection in the Dispose step.  If <see cref="IManagedConnection"/> opened the connection itself during construction then this flag will default
+        /// to true otherwise it will default to false.
+        /// </summary>
+        bool CloseOnDispose { get; set; }
     }
 }