+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed bug when calling SetDoNotReType multiple times on the same DataColumn
+
 ## [0.10.5] - 2019-09-16
 
 ### Added